% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.rpa.hitchip.R
\name{calculate.rpa}
\alias{calculate.rpa}
\title{RPA with HITChip}
\usage{
calculate.rpa(level, phylo, oligo.data)
}
\arguments{
\item{level}{level}

\item{phylo}{phylo}

\item{oligo.data}{oligo.data}
}
\value{
RPA preprocessed data
}
\description{
Fit RPA for HITChip.
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

