% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.probeset.hitchip.R
\name{get.probeset}
\alias{get.probeset}
\title{Get probeset}
\usage{
get.probeset(name, level, taxonomy, probedata, log10 = TRUE)
}
\arguments{
\item{name}{name}

\item{level}{taxonomic level}

\item{taxonomy}{taxonomy}

\item{probedata}{oligos vs. samples preprocessed data matrix; 
absolute scale}

\item{log10}{Logical. Logarithmize the data TRUE/FALSE}
}
\value{
probeset data matrix
}
\description{
Get probeset matrix.
}
\examples{
  #taxonomy <- GetPhylogeny('HITChip', 'filtered')
  #data.dir <- system.file("extdata", package = "microbiome")
  #probedata <- read_hitchip(data.dir, "rpa")$probedata
  #ps <- get.probeset('Akkermansia', 'L2', taxonomy, probedata)
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
