\name{probe.parameters.tolist}
\alias{probe.parameters.tolist}
\title{probe.parameters.tolist}
\usage{
  probe.parameters.tolist(probe.parameters)
}
\arguments{
  \item{probe.parameters}{A data.frame with alpha, betas,
  tau2, affinities, quantile.basis}
}
\description{
  Convert probe parameter table into a list format
}
\examples{
# df <- probe.parameters.tolist(probe.parameters.table)
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("RPA")
}
\keyword{utilities}

