\name{probetable}
\alias{probetable}
\title{probetable}
\usage{
  probetable(probe.parameters)
}
\arguments{
  \item{probe.parameters}{A list with alpha, betas,
  variances and affinities}
}
\description{
  Convert probe-level hyperparameter lists into a table
  format.
}
\examples{
# df <- probetable(probe.parameters)
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("RPA")
}
\keyword{utilities}

