% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.probesets.hitchip.R
\name{retrieve.probesets}
\alias{retrieve.probesets}
\title{Retrieve probesets}
\usage{
retrieve.probesets(tax.table, level = "species", name = NULL)
}
\arguments{
\item{tax.table}{data.frame with oligo - phylotype mapping info}

\item{level}{phylotype level for probesets}

\item{name}{specify phylotypes to check (optional)}
}
\value{
A list. Probes for each phylotype.
}
\description{
List probes for each probeset in taxonomic data.
}
\examples{
  #tax.table <- GetPhylogeny('HITChip')
  #sets <- retrieve.probesets(tax.table, 'species', 'Weissella confusa')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{internal}

