% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.performance.R
\name{collect.hyperparameters}
\alias{collect.hyperparameters}
\title{collect.hyperparameters}
\usage{
collect.hyperparameters(
  batches,
  unique.run.identifier,
  save.batches.dir,
  save.batches,
  verbose = TRUE
)
}
\arguments{
\item{batches}{batch list}

\item{unique.run.identifier}{Batch file identifier string}

\item{save.batches.dir}{Batch file directory}

\item{save.batches}{Logical. Determines whether batches are available.}

\item{verbose}{verbose}
}
\description{
Collect probe-level parameters during online-learning from the batch files.
}
\examples{
# hpe <- collect.hyperparameters(batches, unique.run.identifier, save.batches.dir, save.batches)
}
\references{
See citation("RPA")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
