% Generated by roxygen2 (4.0.1): do not edit by hand
\name{probe.performance}
\alias{probe.performance}
\title{probe.performance}
\usage{
probe.performance(probe.parameters, abatch, sets = NULL)
}
\arguments{
\item{probe.parameters}{List with affinities and variances for the probesets}

\item{abatch}{Affybatch used in the analysis}

\item{sets}{Specify the probesets to include in the output. Default: All probesets}
}
\value{
Data frame of probe-level parameter estimates
}
\description{
Provide a table of probe-level parameter estimates (affinity and stochastic noise) for RPA output.
}
\examples{
# library(affydata) data(Dilution); rpa.results <- rpa.complete(Dilution); tab <- probe.parameters(rpa.results); df <- df[order(abs(df$tau2), decreasing = TRUE),]
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("RPA")
}
\keyword{utilities}

