% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.performance.R
\name{probetable}
\alias{probetable}
\title{probetable}
\usage{
probetable(probe.parameters)
}
\arguments{
\item{probe.parameters}{A list with alpha, betas, variances and affinities}
}
\description{
Convert probe-level hyperparameter lists into a table format.
}
\examples{
# df <- probetable(probe.parameters)
}
\references{
See citation("RPA")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
