% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPA-package.R
\docType{package}
\name{RPA-package}
\alias{RPA-package}
\alias{RPA}
\title{RPA: probabilistic analysis of probe reliability and gene expression}
\description{
Brief summary of the RPA package
}
\details{
\tabular{ll}{
Package: \tab RPA\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2008-2016\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}
}
\examples{
#
}
\references{
See citation("RPA")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{package}
