\docType{package}
\name{RPA-package}
\alias{RPA}
\alias{RPA-package}
\title{RPA: probabilistic analysis of probe reliability and gene expression}
\description{
  Brief summary of the RPA package
}
\details{
  \tabular{ll}{ Package: \tab RPA\cr Type: \tab Package\cr
  Version: \tab See sessionInfo() or DESCRIPTION file\cr
  Date: \tab 2012\cr License: \tab FreeBSD\cr LazyLoad:
  \tab yes\cr }
}
\examples{
#
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("RPA")
}
\keyword{package}

