% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpa.plot.R
\name{rpa.plot}
\alias{rpa.plot}
\title{rpa.plot}
\usage{
rpa.plot(
  x,
  set,
  highlight.probes = NULL,
  pcol = "darkgrey",
  mucol = "black",
  ecol = "red",
  external.signal = NULL,
  main = NULL,
  plots = "all",
  ...
)
}
\arguments{
\item{x}{Output from rpa.complete function}

\item{set}{probeset}

\item{highlight.probes}{mark probes for highlight}

\item{pcol}{probe color}

\item{mucol}{probeset signal color}

\item{ecol}{external signal color}

\item{external.signal}{external signal to be plotted on top}

\item{main}{title}

\item{plots}{plot type}

\item{...}{other arguments to be passed}
}
\value{
Used for its side-effects. Returns probes x samples matrix of probe-level data plotted on the image.
}
\description{
Plot RPA results and probe-level data for a specified probeset.
}
\details{
Plots the preprocessed probe-level observations, estimated probeset-level signal, and probe-specific variances. It is also possible to highlight individual probes and external summary measures.
}
\examples{
# 
}
\references{
See citation("RPA")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{methods}
