% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.hyperparameters.R
\name{updating.hyperparameters}
\alias{updating.hyperparameters}
\title{updating hyperparameters}
\usage{
updating.hyperparameters(
  q,
  set.inds,
  verbose,
  mc.cores = 1,
  alpha,
  betas,
  epsilon
)
}
\arguments{
\item{q}{probes x samples matrix}

\item{set.inds}{Probe set indices}

\item{verbose}{Print progress information}

\item{mc.cores}{Number of cores for parallel computation}

\item{alpha}{alpha hyperparameter}

\item{betas}{beta hyperparameters}

\item{epsilon}{Convergence parameter}
}
\value{
List with the following elements: alpha, betas, s2s (variances)
}
\description{
Hyperparameter update.
}
\examples{
# 
}
\references{
See citation("RPA")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
